spaceai/requirements.txt
Django==4.2
uvicorn 
gunicorn
whitenoise[brotli]
django-compressor
django-compression-middleware
django-redis
pillow
cssmin
jsmin